backend/requirements.txt
fastapi
uvicorn
google-cloud-aiplatform
google-api-python-client
python-multipart
icalendar
pydantic-settings
python-dotenv
